fix: PR Fixes - rename ResultProvider to VariantsResultProvider
EX-6762
